Add TextProcessor
Needed to manage xplanner:text. This attribute is useful to create shortcuts in
any page. So you can access to the link pressing the ALT key and the key set in
the label using the & character. An example:

	This is a &label

Will be displayed as:

	This a label

And set an accesskey to L and a title to `ALT+L`. Pressing the combination of
ALT and L, you can access the link.